Added overload for `&` and `|` in the rules close #361 (#363)
