+++ vkbottle/dispatch/__init__.py
@@ -1,9 +1,8 @@
 from .abc import ABCRouter
 from .bot_router import BotRouter
 from .dispenser import ABCStateDispenser, BuiltinStateDispenser
-from .filter import ABCFilter, AndFilter, OrFilter
 from .handlers import ABCHandler
 from .middlewares import BaseMiddleware, MiddlewareResponse
 from .return_manager import BaseReturnManager
-from .rules import ABCRule
+from .rules import ABCRule, ABCFilter, AndFilter, OrFilter
 from .views import ABCView, ABCDispenseView, ABCMessageView, MessageView, RawEventView
