--- vkbottle/dispatch/filter/__init__.py
@@ -1,2 +1,0 @@
-from .abc import ABCFilter
-from .builtin import AndFilter, OrFilter
