+++ vkbottle/dispatch/rules/__init__.py
@@ -1,1 +1,1 @@
-from .abc import ABCRule
+from .abc import ABCFilter, ABCRule, AndFilter, OrFilter
